major: storage layer is now complete after the implementation of table heap         - Implemented table heap                 - Overflow chaining and directory page chaining implemented fix: fixed page allocation and corruption bugs.

+++ src/error.rs
@@ -16,10 +16,10 @@ pub enum BasaltError {
     CorruptedPage,
 
     #[error("Slot Index {0} is out of bounds")]
-    SlotOutOfBounds(u16),
+    SlotOutOfBounds(u32),
 
     #[error("Slot {0} is a tombstone slot.")]
-    TombstoneSlot(u16),
+    TombstoneSlot(u32),
 
     #[error("The database engine ran out of allocated memory!")]
     OutOfMemory,
